requirements.txt
Django>=4.2
psycopg2-binary>=2.8
djangorestframework==3.15.2
